Add admin credential management features and refactor related components
