+++ backend/src/main/java/com/teamfoundry/backend/account/controller/CompanyCredentialController.java
@@ -1,11 +1,12 @@
 package com.teamfoundry.backend.account.controller;
 
-import com.teamfoundry.backend.account.dto.CompanyCredentialResponse;
+import com.teamfoundry.backend.account.dto.credentials.CompanyApprovalRequest;
+import com.teamfoundry.backend.account.dto.credentials.CompanyCredentialResponse;
 import com.teamfoundry.backend.account.service.CompanyCredentialService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -19,5 +20,19 @@ public class CompanyCredentialController {
     @GetMapping("/companies")
     public List<CompanyCredentialResponse> listPendingCompanyCredentials() {
         return companyCredentialService.listPendingCompanyCredentials();
+    }
+
+    @PostMapping("/companies/{id}/approve")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void approveCompany(@PathVariable int id,
+                               @Valid @RequestBody CompanyApprovalRequest request) {
+        companyCredentialService.approveCompanyCredential(id, request.superAdminPassword());
+    }
+
+    @PostMapping("/companies/{id}/reject")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void rejectCompany(@PathVariable int id,
+                              @Valid @RequestBody CompanyApprovalRequest request) {
+        companyCredentialService.rejectCompanyCredential(id, request.superAdminPassword());
     }
 }
