+++ backend/src/main/java/com/teamfoundry/backend/account/dto/credentials/AdminCredentialUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.teamfoundry.backend.account.dto;
+package com.teamfoundry.backend.account.dto.credentials;
 
 import com.teamfoundry.backend.account.enums.UserType;
 import jakarta.validation.constraints.NotBlank;
@@ -18,5 +18,9 @@ public record AdminCredentialUpdateRequest(
         UserType role,
 
         @Size(min = 8, max = 120, message = "password deve possuir entre 8 e 120 caracteres")
-        String password
+        String password,
+
+        @NotBlank(message = "superAdminPassword é obrigatório")
+        String superAdminPassword
 ) {}
+
