+++ backend/src/main/java/com/teamfoundry/backend/account/repository/CompanyAccountRepository.java
@@ -1,6 +1,6 @@
 package com.teamfoundry.backend.account.repository;
 
-import com.teamfoundry.backend.account.dto.CompanyCredentialResponse;
+import com.teamfoundry.backend.account.dto.credentials.CompanyCredentialResponse;
 import com.teamfoundry.backend.account.model.CompanyAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -20,7 +20,7 @@ public interface CompanyAccountRepository extends JpaRepository<CompanyAccount, 
      * carregamentos adicionais na camada de serviço.
      */
     @Query("""
-            SELECT new com.teamfoundry.backend.account.dto.CompanyCredentialResponse(
+            SELECT new com.teamfoundry.backend.account.dto.credentials.CompanyCredentialResponse(
                 c.id,
                 c.name,
                 c.email,
