+++ backend/src/main/java/com/teamfoundry/backend/security/service/AccountDetailsService.java
@@ -1,7 +1,9 @@
 package com.teamfoundry.backend.security.service;
 
 import com.teamfoundry.backend.account.model.Account;
+import com.teamfoundry.backend.account.model.AdminAccount;
 import com.teamfoundry.backend.account.repository.AccountRepository;
+import com.teamfoundry.backend.account.repository.AdminAccountRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -16,14 +18,28 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class AccountDetailsService implements UserDetailsService {
+
+    private static final String ADMIN_PREFIX = "admin:";
+
     private final AccountRepository repository;
+    private final AdminAccountRepository adminAccountRepository;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        if (username != null && username.startsWith(ADMIN_PREFIX)) {
+            String adminUsername = username.substring(ADMIN_PREFIX.length());
+            AdminAccount admin = adminAccountRepository.findByUsernameIgnoreCase(adminUsername)
+                    .orElseThrow(() -> new UsernameNotFoundException("Admin not found"));
+
+            List<GrantedAuthority> authorities =
+                    List.of(new SimpleGrantedAuthority("ROLE_" + admin.getRole().name()));
+            return new User(username, admin.getPassword(), true, true, true, true, authorities);
+        }
+
         Account acc = repository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        List<GrantedAuthority> auth = List.of(new SimpleGrantedAuthority("ROLE_" + acc.getRole().name()));
-        return new User(acc.getEmail(), acc.getPassword(), acc.isActive(), true, true, true, auth);
+        List<GrantedAuthority> authorities =
+                List.of(new SimpleGrantedAuthority("ROLE_" + acc.getRole().name()));
+        return new User(acc.getEmail(), acc.getPassword(), acc.isActive(), true, true, true, authorities);
     }
 }
-
